Посылаю после дедлайна, но до начала разбора на лекции. Тут плохо, я не все пнял, поэтому не все успел. Но я хочу разобраться и сделать!
